ds4rs_main_app/requirements.txt
altair==4.2.2
amqp==5.1.1
anyio==3.6.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asgiref==3.6.0
asttokens==2.2.1
attrs==22.2.0
backcall==0.2.0
beautifulsoup4==4.12.0
billiard==3.6.4.0
bleach==6.0.0
bqplot==0.12.37
branca==0.6.0
cachetools==5.3.0
celery==5.2.7
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.6
colour==0.1.5
comm==0.1.3
contourpy==1.0.7
cycler==0.11.0
debugpy==1.6.6
decorator==5.1.1
defusedxml==0.7.1
Django==4.1.7
duckdb==0.7.1
earthengine-api==0.1.347
ee-extra==0.0.15
eerepr==0.0.4
entrypoints==0.4
executing==1.2.0
fastjsonschema==2.16.3
ffmpeg-python==0.2.0
filelock==3.10.7
folium==0.14.0
fonttools==4.39.3
fqdn==1.5.1
future==0.18.3
GDAL @ file:///D:/3.%20Projects/dash/GDAL-3.2.3-cp39-cp39-win_amd64.whl
gdown==4.7.1
geeadd==0.5.6
geemap==0.20.4
geocoder==1.38.1
geojson==3.0.1
google-api-core==2.11.0
google-api-python-client==2.83.0
google-auth==2.17.1
google-auth-httplib2==0.1.0
google-cloud-core==2.3.2
google-cloud-storage==2.8.0
google-crc32c==1.5.0
google-resumable-media==2.4.1
googleapis-common-protos==1.59.0
httplib2==0.22.0
idna==3.4
importlib-metadata==6.1.0
importlib-resources==5.12.0
ipyevents==2.0.1
ipyfilechooser==0.6.0
ipykernel==6.22.0
ipyleaflet==0.17.2
ipython==8.12.0
ipython-genutils==0.2.0
ipytree==0.2.2
ipywidgets==7.7.5
isoduration==20.11.0
jedi==0.18.2
Jinja2==3.1.2
jsonpointer==2.3
jsonschema==4.17.3
jupyter-events==0.6.3
jupyter_client==8.1.0
jupyter_core==5.3.0
jupyter_server==2.5.0
jupyter_server_terminals==0.4.4
jupyterlab-pygments==0.2.2
jupyterlab-widgets==1.1.4
kiwisolver==1.4.4
kombu==5.2.4
logzero==1.7.0
MarkupSafe==2.1.2
matplotlib==3.7.1
matplotlib-inline==0.1.6
mistune==2.0.5
nbclassic==0.5.4
nbclient==0.7.2
nbconvert==7.2.10
nbformat==5.8.0
nest-asyncio==1.5.6
notebook==6.5.3
notebook_shim==0.2.2
numpy==1.24.2
packaging==23.0
pandas==1.5.3
pandocfilters==1.5.0
parso==0.8.3
pickleshare==0.7.5
Pillow==9.4.0
platformdirs==3.2.0
plotly==5.14.0
prometheus-client==0.16.0
prompt-toolkit==3.0.38
protobuf==4.22.1
psutil==5.9.4
psycopg2-binary==2.9.5
pure-eval==0.2.2
pyarrow==11.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyCRS==1.0.2
Pygments==2.14.0
pylance==0.4.1
pyparsing==3.0.9
pyperclip==1.8.2
pyrsistent==0.19.3
pyshp==2.3.1
PySocks==1.7.1
python-box==7.0.1
python-dateutil==2.8.2
python-json-logger==2.0.7
pytz==2023.3
pywin32==306
pywinpty==2.0.10
PyYAML==6.0
pyzmq==25.0.2
ratelim==0.1.6
requests==2.28.2
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rsa==4.9
sankee==0.2.3
scooby==0.7.1
Send2Trash==1.8.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.4
sqlparse==0.4.3
stack-data==0.6.2
tenacity==8.2.2
terminado==0.17.1
tinycss2==1.2.1
toolz==0.12.0
tornado==6.2
tqdm==4.65.0
traitlets==5.9.0
traittypes==0.2.1
typing_extensions==4.5.0
tzdata==2023.3
uri-template==1.2.0
uritemplate==4.1.1
urllib3==1.26.15
vine==5.0.0
wcwidth==0.2.6
webcolors==1.13
webencodings==0.5.1
websocket-client==1.5.1
whitebox==2.3.1
whiteboxgui==2.3.0
widgetsnbextension==3.6.4
xyzservices==2023.2.0
zipp==3.15.0
django-environ==0.9.0
gunicorn==20.1.0